rename system solvers
