+++ RefraSin.TEPSolver/ISolverSession.cs
@@ -1,8 +1,8 @@
 using Microsoft.Extensions.Logging;
 using RefraSin.MaterialData;
+using RefraSin.TEPSolver.RootFinding;
 using RefraSin.TEPSolver.StepValidators;
 using RefraSin.TEPSolver.StepVectors;
-using RefraSin.TEPSolver.SystemSolvers;
 using RefraSin.TEPSolver.TimeSteppers;
 using Node = RefraSin.TEPSolver.ParticleModel.Node;
 using Particle = RefraSin.TEPSolver.ParticleModel.Particle;
@@ -77,5 +77,5 @@ public interface ISolverSession
 
     public IReadOnlyList<IStepValidator> StepValidators { get; }
 
-    public ISystemSolver SystemSolver { get; }
+    public IRootFinder RootFinder { get; }
 }
