+++ RefraSin.TEPSolver/RootFinding/IRootFinder.cs
@@ -1,0 +1,8 @@
+using RefraSin.TEPSolver.StepVectors;
+
+namespace RefraSin.TEPSolver.RootFinding;
+
+public interface IRootFinder
+{
+    public StepVector FindRoot(ISolverSession solverSession, StepVector initialGuess);
+}
