--- RefraSin.TEPSolver/SystemSolvers/ISystemSolver.cs
@@ -1,8 +1,0 @@
-using RefraSin.TEPSolver.StepVectors;
-
-namespace RefraSin.TEPSolver.SystemSolvers;
-
-public interface ISystemSolver
-{
-    public StepVector Solve(ISolverSession solverSession, StepVector initialGuess);
-}
