+++ RefraSin.TEPSolver/TimeSteppers/AdamsMoultonTimeStepper.cs
@@ -7,7 +7,7 @@ internal class AdamsMoultonTimeStepper : ITimeStepper
     /// <inheritdoc />
     public StepVector Step(ISolverSession solverSession, StepVector initialGuess)
     {
-        var step = solverSession.SystemSolver.Solve(solverSession, initialGuess);
+        var step = solverSession.RootFinder.FindRoot(solverSession, initialGuess);
 
         if (solverSession.LastStep is not null)
             return (step + solverSession.LastStep) / 2;
